# Inline the `isCmd` check in the `Parser.shift` method