@timvandermeij 2019-07-22
Nice optimization!